TallyAutoPilot/requirements.txt
streamlit
pandas
requests
PyYAML
openai
pdf2image
pytesseract
Pillow